refactor: Rename store folder
