+++ src/components/navbars/ResultsNavbar.tsx
