+++ vite.config.ts
@@ -15,6 +15,7 @@ export default defineConfig({
       "@shared": path.resolve(__dirname, "src/shared"),
       "@components": path.resolve(__dirname, "src/components"),
       "@utils": path.resolve(__dirname, "src/utils"),
+      "@services": path.resolve(__dirname, "src/services"),
     },
   },
 });
